[libgctools] Checksum::AddInvDual16(): Cast siz to int before using `-`.
MSVC doesn't like this: (AppVeyor, MSVC 2013)

src\libgctools\Checksum.cpp(74): error C4146: unary minus operator applied to unsigned type, result still unsigned

+++ src/libgctools/Checksum.cpp
@@ -71,7 +71,7 @@ uint32_t AddInvDual16(const uint16_t *buf, uint32_t siz, ChkEndian endian)
 
 	// NOTE: Integer overflow/underflow is expected here.
 	uint16_t chk1 = 0;
-	uint16_t chk2 = -siz;
+	uint16_t chk2 = (uint16_t)(-(int)siz);
 
 	if (endian != CHKENDIAN_LITTLE) {
 		// Big-endian system. (PowerPC, etc.)
